TVB-2775. Extracted from FileHelper the methods that are used only from the REST layer in order to keep it more lightweight.
