+++ framework_tvb/tvb/core/entities/file/files_helper.py
@@ -34,16 +34,15 @@
 
 import os
 import shutil
-import tempfile
 from threading import Lock
 from zipfile import ZipFile, ZIP_DEFLATED, BadZipfile
+
 from tvb.basic.logger.builder import get_logger
 from tvb.basic.profile import TvbProfile
 from tvb.core.decorators import synchronized
 from tvb.core.entities.file.exceptions import FileStructureException
 from tvb.core.entities.file.xml_metadata_handlers import XMLReader, XMLWriter
 from tvb.core.entities.transient.structure_entities import GenericMetaData
-from werkzeug.utils import secure_filename
 
 LOCK_CREATE_FOLDER = Lock()
 
@@ -418,23 +417,6 @@ class FilesHelper(object):
         if os.path.isfile(file_path):
             return int(os.path.getsize(file_path) / 1024)
         return 0
-
-    @staticmethod
-    def save_temporary_file(file, destination_folder=None):
-        filename = secure_filename(file.filename)
-        if destination_folder is None:
-            destination_folder = FilesHelper.create_temp_folder()
-        full_path = os.path.join(destination_folder, filename)
-        file.save(full_path)
-
-        return full_path
-
-    @staticmethod
-    def create_temp_folder():
-        temp_name = tempfile.mkdtemp(dir=TvbProfile.current.TVB_TEMP_FOLDER)
-        folder = os.path.join(TvbProfile.current.TVB_TEMP_FOLDER, temp_name)
-
-        return folder
 
 
 class TvbZip(ZipFile):
