+++ framework_tvb/tvb/interfaces/rest/client/simulator/simulation_api.py
@@ -30,19 +30,21 @@
 
 import os
 import shutil
+
 from tvb.core.entities.file.files_helper import FilesHelper
 from tvb.core.neocom import h5
 from tvb.interfaces.rest.client.client_decorators import handle_response
 from tvb.interfaces.rest.client.main_api import MainApi
 from tvb.interfaces.rest.commons.strings import RequestFileKey
 from tvb.interfaces.rest.commons.strings import RestLink, LinkPlaceholder
+from tvb.interfaces.rest.commons.files_helper import create_temp_folder
 
 
 class SimulationApi(MainApi):
 
     @handle_response
     def fire_simulation(self, project_gid, session_stored_simulator, temp_folder):
-        temporary_folder = FilesHelper.create_temp_folder()
+        temporary_folder = create_temp_folder()
 
         h5.store_view_model(session_stored_simulator, temporary_folder)
         zip_folder_path = os.path.join(temp_folder, RequestFileKey.SIMULATION_FILE_NAME.value)
